requirements.txt
# PyTorch with CUDA support (install separately using pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121)
# For CUDA 12.1+ (compatible with CUDA 13.0)
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
pyyaml>=6.0
pandas>=2.0.0
scipy>=1.10.0
scikit-image>=0.21.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
